+++ pkg/helm/helm_helpers.go
@@ -22,9 +22,9 @@ import (
 
 	"github.com/jenkins-x/jx/pkg/kube"
 	"github.com/jenkins-x/jx/pkg/log"
+	"github.com/jenkins-x/jx/pkg/secreturl"
 	"github.com/jenkins-x/jx/pkg/table"
 	"github.com/jenkins-x/jx/pkg/util"
-	"github.com/jenkins-x/jx/pkg/vaulturl"
 	"github.com/jenkins-x/jx/pkg/version"
 
 	"k8s.io/client-go/kubernetes"
@@ -517,7 +517,7 @@ type InstallChartOptions struct {
 // respecting the installTimeout, looking up or updating Vault with the username and password for the repo.
 // If vaultClient is nil then username and passwords for repos will not be looked up in Vault.
 func InstallFromChartOptions(options InstallChartOptions, helmer Helmer, kubeClient kubernetes.Interface,
-	installTimeout string, vaultClient vaulturl.Client) error {
+	installTimeout string, vaultClient secreturl.Client) error {
 	chart := options.Chart
 	if options.Version == "" {
 		versionsDir := options.VersionsDir
@@ -573,7 +573,7 @@ type HelmRepoCredential struct {
 
 // DecorateWithSecrets will replace any vault: URIs with the secret from vault. Safe to call with a nil client (
 // no replacement will take place).
-func DecorateWithSecrets(options *InstallChartOptions, vaultClient vaulturl.Client) (func(), error) {
+func DecorateWithSecrets(options *InstallChartOptions, vaultClient secreturl.Client) (func(), error) {
 	cleanup := func() {
 	}
 	if vaultClient != nil {
@@ -595,7 +595,7 @@ func DecorateWithSecrets(options *InstallChartOptions, vaultClient vaulturl.Clie
 			if err != nil {
 				return cleanup, errors.Wrapf(err, "reading file %s", valueFile)
 			}
-			newValues, err := vaulturl.ReplaceURIs(string(bytes), vaultClient)
+			newValues, err := secreturl.ReplaceURIs(string(bytes), vaultClient)
 			if err != nil {
 				return cleanup, errors.Wrapf(err, "replacing vault URIs")
 			}
@@ -615,7 +615,7 @@ func DecorateWithSecrets(options *InstallChartOptions, vaultClient vaulturl.Clie
 // The repo name may have a suffix added in order to prevent name collisions, and is returned for this reason.
 // The username and password will be stored in vault for the URL (if vault is enabled).
 func AddHelmRepoIfMissing(helmURL, repoName, username, password string, helmer Helmer,
-	vaultClient vaulturl.Client, in terminal.FileReader,
+	vaultClient secreturl.Client, in terminal.FileReader,
 	out terminal.FileWriter, outErr io.Writer) (string, error) {
 	missing, existingName, err := helmer.IsRepoMissing(helmURL)
 	if err != nil {
@@ -662,7 +662,7 @@ func AddHelmRepoIfMissing(helmURL, repoName, username, password string, helmer H
 }
 
 // DecorateWithCredentials will, if vault is installed, store or replace the username or password
-func DecorateWithCredentials(repo string, username string, password string, vaultClient vaulturl.Client, in terminal.FileReader,
+func DecorateWithCredentials(repo string, username string, password string, vaultClient secreturl.Client, in terminal.FileReader,
 	out terminal.FileWriter, outErr io.Writer) (string,
 	string, error) {
 	if repo != "" && vaultClient != nil {
