+++ pkg/vault/helpers_test.go
@@ -5,7 +5,7 @@ import (
 	"io/ioutil"
 	"testing"
 
-	"github.com/jenkins-x/jx/pkg/vaulturl"
+	"github.com/jenkins-x/jx/pkg/secreturl"
 	"github.com/pborman/uuid"
 
 	"github.com/stretchr/testify/assert"
@@ -35,7 +35,7 @@ func TestReplaceURIs(t *testing.T) {
 	pegomock.When(vaultClient.Read(pegomock.EqString(path))).ThenReturn(map[string]interface{}{
 		key: secret,
 	}, nil)
-	result, err := vaulturl.ReplaceURIs(valuesyaml, vaultClient)
+	result, err := secreturl.ReplaceURIs(valuesyaml, vaultClient)
 	assert.NoError(t, err)
 	assert.NoError(t, err)
 	assert.Equal(t, fmt.Sprintf(`foo:
@@ -63,7 +63,7 @@ func TestReplaceRealExampleURI(t *testing.T) {
 	pegomock.When(vaultClient.Read(pegomock.EqString(path))).ThenReturn(map[string]interface{}{
 		key: secret,
 	}, nil)
-	result, err := vaulturl.ReplaceURIs(valuesyaml, vaultClient)
+	result, err := secreturl.ReplaceURIs(valuesyaml, vaultClient)
 	assert.NoError(t, err)
 	assert.NoError(t, err)
 	assert.Equal(t, fmt.Sprintf(`foo:
